# backend-go/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Xianzhi AI User Core API",
    "version": "0.1.0",
    "description": "Initial OpenAPI contract for the shared user clients and business SDK."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "loginByPassword",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/auth/wechat-mini-program/login": {
      "post": {
        "operationId": "loginByWechatMiniProgram",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "registerByInvite",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterByInviteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refreshAuthSession",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/user/dashboard": {
      "get": {
        "operationId": "getUserDashboard",
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserDashboardResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/generation-tasks": {
      "get": {
        "operationId": "listGenerationTasks",
        "responses": {
          "200": {
            "description": "Generation task list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenerationTask"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "operationId": "createGenerationTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGenerationTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenerationTaskResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/ppt/generate": {
      "post": {
        "operationId": "createPptGeneration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenerationTaskResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/assets": {
      "get": {
        "operationId": "listAssets",
        "responses": {
          "200": {
            "description": "Asset list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/models": {
      "get": {
        "operationId": "listModels",
        "responses": {
          "200": {
            "description": "Model list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelInfo"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/plans": {
      "get": {
        "operationId": "listPlans",
        "responses": {
          "200": {
            "description": "Commerce plan list.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommercePlan"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/CommercePlanItems"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/member/profile": {
      "get": {
        "operationId": "getMemberProfile",
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemberProfileResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/member/wallet": {
      "get": {
        "operationId": "getMemberWallet",
        "responses": {
          "200": {
            "$ref": "#/components/responses/RoleWalletResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/points/account": {
      "get": {
        "operationId": "getPointsAccount",
        "responses": {
          "200": {
            "$ref": "#/components/responses/PointAccountResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/channel/me": {
      "get": {
        "operationId": "getChannelCenter",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ChannelCenterResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/operation-center/profile": {
      "get": {
        "operationId": "getOperationProfile",
        "responses": {
          "200": {
            "$ref": "#/components/responses/OperationProfileResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/operation-center/agents": {
      "get": {
        "operationId": "listOperationAgents",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ItemsResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/operation-center/orders": {
      "get": {
        "operationId": "listOperationOrders",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ItemsResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/operation-center/commissions": {
      "get": {
        "operationId": "listOperationCommissions",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ItemsResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    },
    "responses": {
      "AuthResponse": {
        "description": "Authenticated user session.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthResponse"
            }
          }
        }
      },
      "GenerationTaskResponse": {
        "description": "Generation task.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenerationTask"
            }
          }
        }
      },
      "UserDashboardResponse": {
        "description": "User dashboard data.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserDashboardResponse"
            }
          }
        }
      },
...
      "MemberProfileResponse": {
        "description": "Member profile data.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MemberProfileResponse"
            }
          }
        }
      },
...
      "OkResponse": {
        "description": "Operation result.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OkResponse"
            }
          }
        }
      },
      "ErrorResponse": {
        "description": "API error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "AnyRecord": {
        "type": "object",
        "additionalProperties": true
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "WechatLoginRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "RegisterByInviteRequest": {
        "type": "object",
        "required": ["username", "email", "password", "confirmPassword", "inviteCode"],
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          },
          "inviteCode": {
            "type": "string"
          }
        }
      },
...
      "GenerationTaskType": {
        "type": "string"
      },
...
      "GenerationTask": {
        "type": "object",
        "required": ["id", "type", "status", "prompt", "model", "pointCost"],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/GenerationTaskType"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "prompt": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "pointCost": {
            "type": "integer"
          },
          "resultIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "params": {
            "$ref": "#/components/schemas/AnyRecord"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "workerFinishedAt": {
            "type": "string"
          }
        }
      },
      "CreateGenerationTaskRequest": {
        "type": "object",
        "required": ["type", "prompt", "model"],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/GenerationTaskType"
          },
          "module_code": {
            "type": "string"
          },
          "moduleCode": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "params": {
            "$ref": "#/components/schemas/AnyRecord"
          }
        }
      },
...
      "CommercePlanItems": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommercePlan"
            }
          }
        }
      },
...
    }
  }
}